+++ src/main/java/io/github/tehstoneman/betterstorage/client/gui/GuiCardboardBox.java
@@ -41,6 +41,7 @@ public class GuiCardboardBox extends ContainerScreen< ContainerCardboardBox >
 	@Override
 	public void render( int mouseX, int mouseY, float partialTicks )
 	{
+		renderBackground();
 		super.render( mouseX, mouseY, partialTicks );
 		renderHoveredToolTip( mouseX, mouseY );
 	}