+++ src/main/java/io/github/tehstoneman/betterstorage/client/gui/GuiKeyring.java
@@ -19,6 +19,7 @@ public class GuiKeyring extends ContainerScreen< ContainerKeyring >
 	@Override
 	public void render( int mouseX, int mouseY, float partialTicks )
 	{
+		renderBackground();
 		super.render( mouseX, mouseY, partialTicks );
 		renderHoveredToolTip( mouseX, mouseY );
 	}